requirements.txt
Flask==2.0.3
docx2pdf==0.1.8
matplotlib==3.5.1
matplotlib-inline==0.1.3
opencv-python==4.5.5
Pillow==9.0.1
pytesseract==0.3.9
python-docx==0.8.11
requests==2.27.1
Werkzeug==2.0.3



